+++ app/src/main/java/com/inky/fitnesscalendar/ui/components/BaseEditDialog.kt
@@ -17,6 +17,7 @@ import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.testTag
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
@@ -90,7 +91,8 @@ fun ColumnScope.OkayCancelRow(
         }
         TextButton(
             onClick = onSave,
-            enabled = saveEnabled
+            enabled = saveEnabled,
+            modifier = Modifier.testTag("button-confirm")
         ) {
             saveText()
         }